init commit
so far, we aren't doing anything too fancy. there's some leftover code
that should be scrubbed from this where i tried making a goroutine to
listen on the client-specific port. with this commit, you can point
quake at localhost and get to the "connecting to server" step (one
packet exchange)