Added initial packets for server-transmitted unreliable messages.
encode.go:
	added a case to Encode, "case *UnreliableMessagePacket: return
	encodeUnreliableMessage"
	use is not at all implemented
packets.go:
	defined the UnreliableMessagePacket interface
	any structs that implement this interface should have an
	AddOpcode method to add to it's opcode messages. This interface
	probably will and probably should change in the future

	GameServerPacket implemented, and uses UnreliableMessage
	interface

	Addopcode appends info into the packet's messages buffer. what's
	added is dependant on the opcode used.

	Sv_TimeStampPacket is an example of a simple packet

	Sv_ClientdataPacket is an example of a packet that uses bitmasks
	to delineate size. It's Build method allows it to append only
	necessary fields. Right now it is only built to give the client
	a rocket launcher. this will eventually interface with the
	QuakeC component I imagine.

most of what I added to packet.go remains untested and largely unfounded
in real go programming (ie, i need to revisit the Messages []uint8 and
see whether a push or append is more applicable, and see if a buffer is
a better choice. also, in general stop being such a bad programmer. i
think it's a decent outline of stuff, next goal is to get it all tested properly